backend/requirements.txt
fastapi==0.116.1
httpx==0.28.1
psycopg[binary]==3.2.10
python-dotenv==1.1.1
sqlalchemy==2.0.43
uvicorn[standard]==0.35.0
